Change how types are adjusted when substituting jokers
Instead of fixing identifier type and binop type in one function,
first fix all identifiers (replace joker identifiers with their
expressions, and find declarations in CFA for other identifiers),
then fix-up type of binops. Fix-up is not clever and simply
picks type of the second operand as binop calculation and result type,
which is not correct but is consistent with previous implementation.